`nodejs/gulp-javascript-webserver` - Created a independent configurer
I created a script, `configurer` to delegate all the configuration
options and required operations to get them to it and uncouple from it
from the `server`. Due it, I created and `index` file which orchestrate
the application cycle, as getting the configuration to after use it to
start the server.

`runDevServer` has been update to use the new `index` script